requirements.txt
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
mutagen==1.47.0
overrides==7.7.0
pyarr==5.2.0
pycryptodomex==3.20.0
PyYAML==6.0.1
requests==2.31.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
urllib3==1.26.18
websockets==12.0
yt-dlp==2024.3.10
fastapi
uvicorn[standard]
python-multipart
fastapi-login
passlib[bcrypt]
python-jose
jinja2
rich